Implement code allowing two game modes: human and bot
The player, no matter a human or a bot, has a chance to select their
identity at the beginning of the game. If the player claims that they
are human, they can select rows and columns very naturally. Otherwise,
they need to choose operators 11-20 for rows and 21-30 for columns.

To-do:
- Add preconditions to changeGrid function so that the human player's
actions can be restricted